requirements.txt
# Core LangChain ecosystem
langchain==0.3.7
langchain-core==0.3.63
langchain-community==0.3.7
langchain-huggingface==0.1.2
langchain-google-genai==2.0.2
langchain-pinecone==0.2.12
langchain-text-splitters==0.3.8

# LangGraph
langgraph==0.2.36
langgraph-checkpoint==2.1.1
langgraph-sdk==0.1.74

# Pinecone
pinecone==7.3.0
pinecone-client==4.1.0

# Embeddings & ML
sentence-transformers==3.2.1
transformers==4.56.2
torch==2.8.0

# Utilities
python-dotenv==1.0.1
docx2txt==0.8
fastapi==0.115.2
uvicorn==0.30.6
tavily-python==0.3.5

# Fix dependency conflicts
langsmith==0.1.147
urllib3==1.26.20
